cleans up integration directory code
